refactor(sessions): extract LLM session helpers
